requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Document processing
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
beautifulsoup4==4.12.2

# Embeddings and vector stores
sentence-transformers==2.2.2
chromadb==0.4.15
numpy==1.24.3

# Retrieval
rank-bm25==0.2.2

# Reranking
transformers==4.35.2
torch==2.1.1

# LLM generation
accelerate==0.24.1

# Evaluation
ragas==0.1.7
datasets==2.14.5

# Experiment tracking
mlflow==2.10.0

# Data processing
pandas==2.1.3

# Logging and utilities
python-dotenv==1.0.0
